requirements.txt
dbt-core==1.4.6  # Pinned for the SQLite adapter
dbt-sqlite==1.4.0
Jinja2>=3.*
pybadges==3.0.0
pyyaml==6.0
sqlglot==16.4.2
typer==0.9.0
